% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotkegg.r
\name{plot_kegg}
\alias{plot_kegg}
\title{A keggmap plot function}
\usage{
plot_kegg(pg)
}
\arguments{
\item{kegg}{pathway Which pathway you want to plot, plot_kegg('ec00010').}
}
\description{
This function allows you to plot kegg pathway based on xml file.
}
\examples{
plot_kegg('ec00020')
}
\keyword{keggplot}
